+++ lib-h3/src/debug/h3_ccu_pll_dump.c
@@ -2,7 +2,7 @@
  * @file h3_ccu.c
  *
  */
-/* Copyright (C) 2018-2019 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2018-2021 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -30,13 +30,13 @@
 
 void __attribute__((cold)) h3_ccu_pll_dump(void) {
 	uart0_puts("PLL (Hz)\n");
-	uart0_printf("CPUX=%ld\n", (long int) h3_ccu_get_pll_rate(CCU_PLL_CPUX));
-	uart0_printf("AUDIO=%ld\n", (long int) h3_ccu_get_pll_rate(CCU_PLL_AUDIO));
-	uart0_printf("VIDEO=%ld\n", (long int) h3_ccu_get_pll_rate(CCU_PLL_VIDEO));
-	uart0_printf("VE=%ld\n", (long int) h3_ccu_get_pll_rate(CCU_PLL_VE));
-	uart0_printf("DDR=%ld\n", (long int) h3_ccu_get_pll_rate(CCU_PLL_DDR));
-	uart0_printf("PERIPH0=%ld\n", (long int) h3_ccu_get_pll_rate(CCU_PLL_PERIPH0));
-	uart0_printf("GPU=%ld\n", (long int) h3_ccu_get_pll_rate(CCU_PLL_GPU));
-	uart0_printf("PERIPH1=%ld\n", (long int) h3_ccu_get_pll_rate(CCU_PLL_PERIPH1));
-	uart0_printf("DE=%ld\n", (long int) h3_ccu_get_pll_rate(CCU_PLL_DE));
+	uart0_printf("CPUX=%u\n", h3_ccu_get_pll_rate(CCU_PLL_CPUX));
+	uart0_printf("AUDIO=%u\n", h3_ccu_get_pll_rate(CCU_PLL_AUDIO));
+	uart0_printf("VIDEO=%u\n", h3_ccu_get_pll_rate(CCU_PLL_VIDEO));
+	uart0_printf("VE=%u\n", h3_ccu_get_pll_rate(CCU_PLL_VE));
+	uart0_printf("DDR=%u\n", h3_ccu_get_pll_rate(CCU_PLL_DDR));
+	uart0_printf("PERIPH0=%u\n", h3_ccu_get_pll_rate(CCU_PLL_PERIPH0));
+	uart0_printf("GPU=%u\n", h3_ccu_get_pll_rate(CCU_PLL_GPU));
+	uart0_printf("PERIPH1=%u\n", h3_ccu_get_pll_rate(CCU_PLL_PERIPH1));
+	uart0_printf("DE=%u\n", h3_ccu_get_pll_rate(CCU_PLL_DE));
 }
