+++ lib-l6470dmx/include/modestore.h
@@ -26,13 +26,13 @@
 #ifndef MODESTORE_H_
 #define MODESTORE_H_
 
-#include <stdint.h>
+#include <cstdint>
 
 class ModeStore {
 public:
 	virtual ~ModeStore() {}
 
-	virtual void SaveDmxStartAddress(uint8_t nMotorIndex, uint16_t nDmxStartAddress)=0;
+	virtual void SaveDmxStartAddress(uint32_t nMotorIndex, uint16_t nDmxStartAddress)=0;
 };
 
 #endif /* MODESTORE_H_ */
