+++ lib-ltc/include/h3/systimereader.h
@@ -26,7 +26,7 @@
 #ifndef H3_SYSTIMEREADER_H_
 #define H3_SYSTIMEREADER_H_
 
-#include <stdint.h>
+#include <cstdint>
 #include <time.h>
 
 #include "ltc.h"
@@ -42,7 +42,7 @@ public:
 	void Print();
 
 	//
-	void HandleRequest(void *pBuffer = nullptr, uint32_t nBufferLength = 0);
+	void HandleRequest(void *pBuffer = nullptr, uint16_t nBufferLength = 0);
 
 	// Control
 	void ActionStart();
